refactor: Remove duplicate code in keyFigureText macro and import from keyFigureText lib
