+++ src/main/resources/lib/ssb/parts/keyFigureText.ts
@@ -9,57 +9,7 @@ import { parseKeyFigure } from '/lib/ssb/parts/keyFigure'
 import { type RequestWithCode } from '/lib/types/municipalities'
 import { type KeyFigure } from '/site/content-types'
 
-function getKeyFigureTextValuesFromString(ingress: string | undefined): KeyFigureTextValues | undefined {
-  if (ingress) {
-    const keyFigureId = ingress.match(/keyFigure="([^"]+)"/)
-    const keyFigure = keyFigureId ? get({ key: keyFigureId[1] }) : undefined
-
-    const municipality = getMunicipality({ code: keyFigure?.data.default } as RequestWithCode)
-    const language = keyFigure?.language ? keyFigure.language : 'nb'
-    const keyFigureData = parseKeyFigure(keyFigure as Content<KeyFigure>, municipality, 'master', language)
-    const sourceText = getKeyFigureSourceText(keyFigure as Content<KeyFigure>)
-
-    const text = ingress.match(/text="([^"]+)"/)
-    const overwriteIncrease = ingress.match(/overwriteIncrease="([^"]+)"/)
-    const overwriteDecrease = ingress.match(/overwriteDecrease="([^"]+)"/)
-    const overwriteNoChange = ingress.match(/overwriteNoChange="([^"]+)"/)
-
-    return {
-      keyFigureData,
-      sourceText,
-      language,
-      text: text ? text[1] : undefined,
-      overwriteIncrease: overwriteIncrease ? overwriteIncrease[1] : undefined,
-      overwriteDecrease: overwriteDecrease ? overwriteDecrease[1] : undefined,
-      overwriteNoChange: overwriteNoChange ? overwriteNoChange[1] : undefined,
-    }
-  }
-}
-
-export function getIngressWithKeyFigureText(ingress: string | undefined) {
-  const keyFigureTextValues = getKeyFigureTextValuesFromString(ingress)
-
-  if (keyFigureTextValues) {
-    const { keyFigureData, text, overwriteIncrease, overwriteDecrease, overwriteNoChange, language, sourceText } =
-      keyFigureTextValues
-
-    const ingressWithKeyFigureText = parseKeyFigureText(
-      keyFigureData as KeyFigureView,
-      { text, overwriteIncrease, overwriteDecrease, overwriteNoChange } as KeyFigureTextContext,
-      language as string,
-      sourceText
-    )
-    return ingress?.replace(/\[keyFigureText\b[^]*?\/\]/g, ingressWithKeyFigureText)
-  }
-}
-
-export const getKeyFigureSourceText = (keyFigure: Content<KeyFigure> | undefined): string | undefined => {
-  return forceArray(keyFigure?.data.source).length
-    ? forceArray(keyFigure?.data.source).map(({ title }) => title)[0]
-    : undefined
-}
-
-export function getLocalizedChangeDirection(
+function getLocalizedChangeDirection(
   changeDirection: KeyFigureChanges['changeDirection'] | undefined,
   language: string,
   overwriteIncrease: string | undefined,
@@ -99,7 +49,7 @@ export function getLocalizedChangeDirection(
   return changeDirection
 }
 
-export function getChangeValue(changes: KeyFigureChanges | undefined): string | undefined {
+function getChangeValue(changes: KeyFigureChanges | undefined): string | undefined {
   let changeValue
   if (changes?.changeText) {
     // When there are no changes, the change period for 'same' is displayed (e.g. "Ingen endring") as the change value. It should be displayed as empty string instead
@@ -113,8 +63,58 @@ export function getChangeValue(changes: KeyFigureChanges | undefined): string | 
 }
 
 // We have to manually strip away 'endring' for some change periods; sometimes a change period is prefixed with 'endring' e.g. 'endring fra året før'
-export const getChangePeriod = (changes: KeyFigureChanges | undefined): string | undefined =>
+const getChangePeriod = (changes: KeyFigureChanges | undefined): string | undefined =>
   changes?.changePeriod ? changes.changePeriod.toLowerCase().replace('endring ', '') : undefined
+
+function getKeyFigureTextValuesFromString(ingress: string | undefined): KeyFigureTextValues | undefined {
+  if (ingress) {
+    const keyFigureId = ingress.match(/keyFigure="([^"]+)"/)
+    const keyFigure = keyFigureId ? get({ key: keyFigureId[1] }) : undefined
+
+    const municipality = getMunicipality({ code: keyFigure?.data.default } as RequestWithCode)
+    const language = keyFigure?.language ? keyFigure.language : 'nb'
+    const keyFigureData = parseKeyFigure(keyFigure as Content<KeyFigure>, municipality, 'master', language)
+    const sourceText = getKeyFigureSourceText(keyFigure as Content<KeyFigure>)
+
+    const text = ingress.match(/text="([^"]+)"/)
+    const overwriteIncrease = ingress.match(/overwriteIncrease="([^"]+)"/)
+    const overwriteDecrease = ingress.match(/overwriteDecrease="([^"]+)"/)
+    const overwriteNoChange = ingress.match(/overwriteNoChange="([^"]+)"/)
+
+    return {
+      keyFigureData,
+      sourceText,
+      language,
+      text: text ? text[1] : undefined,
+      overwriteIncrease: overwriteIncrease ? overwriteIncrease[1] : undefined,
+      overwriteDecrease: overwriteDecrease ? overwriteDecrease[1] : undefined,
+      overwriteNoChange: overwriteNoChange ? overwriteNoChange[1] : undefined,
+    }
+  }
+}
+
+export const getKeyFigureSourceText = (keyFigure: Content<KeyFigure> | undefined): string | undefined => {
+  return forceArray(keyFigure?.data.source).length
+    ? forceArray(keyFigure?.data.source).map(({ title }) => title)[0]
+    : undefined
+}
+
+export function getIngressWithKeyFigureText(ingress: string | undefined) {
+  const keyFigureTextValues = getKeyFigureTextValuesFromString(ingress)
+
+  if (keyFigureTextValues) {
+    const { keyFigureData, text, overwriteIncrease, overwriteDecrease, overwriteNoChange, language, sourceText } =
+      keyFigureTextValues
+
+    const ingressWithKeyFigureText = parseKeyFigureText(
+      keyFigureData as KeyFigureView,
+      { text, overwriteIncrease, overwriteDecrease, overwriteNoChange } as KeyFigureTextContext,
+      language as string,
+      sourceText
+    )
+    return ingress?.replace(/\[keyFigureText\b[^]*?\/\]/g, ingressWithKeyFigureText)
+  }
+}
 
 export function parseKeyFigureText(
   keyFigureData: KeyFigureView,
@@ -125,6 +125,7 @@ export function parseKeyFigureText(
   const { title, time, number, numberDescription, changes } = keyFigureData
   const { overwriteIncrease, overwriteDecrease, overwriteNoChange, text } = keyFigureTextMacroInput ?? {}
 
+  const localizedTime = language !== 'en' && time ? time?.toLowerCase() : time
   const changeDirection = getLocalizedChangeDirection(
     changes?.changeDirection,
     language,
@@ -134,19 +135,18 @@ export function parseKeyFigureText(
   )
   const changeValue = getChangeValue(changes)
   const changePeriod = getChangePeriod(changes)
-  const localizedChangePeriod = language !== 'en' && changePeriod ? changePeriod?.toLowerCase() : changePeriod
 
   return text
     ? text
         .replace(/\$tittel/g, title ?? '<mangler tittel>')
-        .replace(/\$tid/g, time ?? '<mangler tid>')
+        .replace(/\$tid/g, localizedTime ?? '<mangler tid>')
         .replace(/\$tall/g, number ?? '<mangler tall>')
         .replace(/\$benevning/g, numberDescription ?? '<mangler benevning>')
         .replace(/\$endringstekst/g, changeDirection ?? '<mangler endringstekst>')
         .replace(/\$endringstall/g, changeValue ?? '<mangler endringstall>')
-        .replace(/\$endringsperiode/g, localizedChangePeriod ?? '<mangler endringsperiode>')
+        .replace(/\$endringsperiode/g, changePeriod ?? '<mangler endringsperiode>')
         .replace(/\$kildetekst/g, sourceText ?? '<mangler kildetekst>')
-    : [title, time, number, numberDescription, changeDirection, changeValue, localizedChangePeriod, sourceText].join(
+    : [title, localizedTime, number, numberDescription, changeDirection, changeValue, changePeriod, sourceText].join(
         ' '
       )
 }
